Made the base paginator work without the request object.
And extended the base paginator to keep BC compatibility with the paginator component.
